refactor(utils): add getDiffInCalendarMonths ESM
