+++ src/utils/index.ts
@@ -4,4 +4,5 @@ export { default as getDiffInCalendarMinutes } from './getDiffInCalendarMinutes.
 export { default as getDiffInCalendarHours } from './getDiffInCalendarHours.js';
 export { default as getDiffInCalendarDays } from './getDiffInCalendarDays.js';
 export { default as getDiffInCalendarWeeks } from './getDiffInCalendarWeeks.js';
+export { default as getDiffInCalendarMonths } from './getDiffInCalendarMonths.js';
 export { default as getDiffInCalendarYears } from './getDiffInCalendarYears.js';
